(PC-6460) Move tests to the page
