+++ src/libs/adresse/fetchPlaces.ts
@@ -1,3 +1,4 @@
+// eslint-disable-next-line import/no-unresolved
 import { FeatureCollection, Point } from 'geojson'
 
 import { Properties, SuggestedPlace } from './types'
